refactor(entry point): simplify main and binary entry points
