+++ package.json
@@ -5,7 +5,7 @@
   "bin": {
     "contentful-export": "./bin/contentful-export"
   },
-  "main": "index.js",
+  "main": "dist/index.js",
   "engines": {
     "node": ">=4.7"
   },
@@ -61,6 +61,7 @@
     "babel-cli": "^6.3.17",
     "babel-core": "^6.23.1",
     "babel-eslint": "^7.1.1",
+    "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-rewire": "^1.0.0-beta-3",
     "babel-preset-env": "^1.2.1",
     "babel-register": "^6.3.13",
